Cast uint32_t printf arguments to match their %u conversions
The first real compile of the firmware turned up two format mismatches
that static review had missed: `seconds` and `gWakeCount` are uint32_t,
which the xtensa toolchain widens to `long unsigned int`, so `%u` did not
match and -Wformat flagged both.

Both are harmless on a 32-bit target, where the two types are passed
identically through varargs, but they are the last warnings in the file
and they hide the next real one. Cast at the call site, matching the fix
already applied to the pulse-width log line.

Verified by cross-compiling src/main.cpp for esp32dev with the
arduino-esp32 3.3.11 core: all three build environments (esp32dev, usb,
bench) now compile clean under -Wall -Wextra.

+++ src/main.cpp
@@ -166,7 +166,7 @@ void runBenchSweep() {
 
 void sleepFor(uint32_t seconds) {
   if (seconds == 0) seconds = 1;
-  Serial.printf("[wind] sleeping %u s\n", seconds);
+  Serial.printf("[wind] sleeping %u s\n", static_cast<unsigned>(seconds));
   Serial.flush();
 
   gSecondsSinceGoodUpdate += seconds;
@@ -190,7 +190,7 @@ void setup() {
   Serial.begin(115200);
   delay(50);
   ++gWakeCount;
-  Serial.printf("\n[wind] wake #%u\n", gWakeCount);
+  Serial.printf("\n[wind] wake #%u\n", static_cast<unsigned>(gWakeCount));
 
   gpio_deep_sleep_hold_dis();
   gpio_hold_dis(static_cast<gpio_num_t>(config::kServoPowerPin));
